+++ src/components/elements/Icon/index.tsx
@@ -1,0 +1,10 @@
+/**
+ * Component build on top of Fontawesome
+ */
+import * as React from "react";
+
+const Icon = ({ className }: { className: string }) => (
+  <i className={className} aria-hidden="true" />
+);
+
+export default Icon;